pyproject.toml
[project]
name = "sidekick-langgraph-agent"
version = "0.1.0"
description = "LangGraph Sidekick App"
requires-python = ">=3.11"
dependencies = [
    "gradio",
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "langchain-experimental",
    "markdown",
    "python-dotenv",
    "pydantic",
    "playwright",
    "requests",
    "typing-extensions",
    "wikipedia",
    "bcrypt",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
    "pre-commit",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "PIE", "T20", "RET", "SIM", "ARG", "PTH", "PL", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
